requirements.txt
mysql==0.0.3
pickle5==0.0.11
python-dotenv==1.0.0